+++ src/components/empty_prompt/empty_prompt.test.tsx
@@ -9,8 +9,8 @@
 import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test';
-
-import { EuiEmptyPrompt } from './empty_prompt';
+import { EuiEmptyPrompt, PADDING_SIZES } from './empty_prompt';
+import { COLORS } from '../panel/panel';
 
 describe('EuiEmptyPrompt', () => {
   test('is rendered', () => {
@@ -75,6 +75,56 @@ describe('EuiEmptyPrompt', () => {
         const component = render(
           <EuiEmptyPrompt actions={['action1', 'action2']} />
         );
+        expect(component).toMatchSnapshot();
+      });
+    });
+
+    describe('footer', () => {
+      test('renders alone', () => {
+        const component = render(<EuiEmptyPrompt footer="footer" />);
+        expect(component).toMatchSnapshot();
+      });
+    });
+
+    describe('layout', () => {
+      test('renders alone', () => {
+        const component = render(<EuiEmptyPrompt layout="horizontal" />);
+        expect(component).toMatchSnapshot();
+      });
+    });
+
+    describe('paddingSize', () => {
+      PADDING_SIZES.forEach((size) => {
+        it(`${size} is rendered`, () => {
+          const component = render(<EuiEmptyPrompt paddingSize={size} />);
+
+          expect(component).toMatchSnapshot();
+        });
+      });
+    });
+
+    describe('color', () => {
+      COLORS.forEach((color) => {
+        test(`${color} is rendered`, () => {
+          const component = render(<EuiEmptyPrompt color={color} />);
+
+          expect(component).toMatchSnapshot();
+        });
+      });
+    });
+
+    describe('styles', () => {
+      test('are rendered', () => {
+        const component = render(
+          <EuiEmptyPrompt
+            style={{
+              background: 'yellow',
+              minWidth: '200px',
+              maxWidth: '600px',
+            }}
+          />
+        );
+
         expect(component).toMatchSnapshot();
       });
     });