+++ src/components/icon/icon.tsx
@@ -28,7 +28,7 @@ const typeToPathMap = {
   agentApp: 'app_fleet',
   aggregate: 'aggregate',
   alert: 'alert',
-  analyzeEvent: 'analyze_event',
+  analyzeEvent: 'analyzeEvent',
   annotation: 'annotation',
   apmApp: 'app_apm',
   apmTrace: 'apm_trace',
@@ -474,7 +474,7 @@ export const COLORS: NamedColor[] = keysOf(colorToClassMap);
 
 type NamedColor = keyof typeof colorToClassMap;
 
-function isNamedColor(name: string): name is NamedColor {
+export function isNamedColor(name: string): name is NamedColor {
   return colorToClassMap.hasOwnProperty(name);
 }
 
@@ -767,6 +767,7 @@ export class EuiIcon extends PureComponent<EuiIconProps, State> {
           focusable={focusable}
           role="img"
           title={title}
+          data-icon-type={this.state.iconTitle}
           {...titleId}
           {...rest}
           {...hideIconEmpty}